Séparer la question du logement en 2 questions distinctes
